Front: fixed the X not working in the forms by moving the div and span with the X into in the index.html rather than repeating in each form html
