+++ _main_project/a_spa_frontend/static/js/navbar/Modal.js
@@ -2,6 +2,7 @@ export class Modal {
   constructor(buttonId, contentForm) {
     this.button = document.getElementById(buttonId);
     this.modal = document.getElementById("formModal");
+    this.form = document.getElementById("modalContent");
     this.initEventListeners();
     this.contentForm = contentForm;
   }
@@ -10,15 +11,15 @@ export class Modal {
     this.button.addEventListener('click', async(event) =>{
       if (event.target === this.button || this.button.contains(event.target)) {
         this.show();
-        this.modal.innerHTML = await this.contentForm.getHtml();
+        this.form.innerHTML = await this.contentForm.getHtml();
         this.contentForm.afterRender();
       }
     });
 
-    // const span = this.modal.querySelector('.close');
-    // span.addEventListener('click', () => {
-    //   this.hide();
-    // });
+    const span = this.modal.querySelector('.close');
+    span.addEventListener('click', () => {
+      this.hide();
+    });
 
     window.onclick = event => {
       if (event.target === this.modal) {
