+++ _main_project/a_spa_frontend/static/js/navbar/navbar.js
@@ -1,6 +1,6 @@
-import RegisterForm from "../views/RegisterForm.js";
-import LoginForm from "../views/LoginForm.js";
-import UserForm from "../views/UserForm.js";
+import RegisterForm from "../forms/RegisterForm.js";
+import LoginForm from "../forms/LoginForm.js";
+import UserForm from "../forms/UserForm.js";
 import { Modal } from "./Modal.js";
 import { navigateTo } from "../index.js";
 import { getProfileInfo } from "../helpers/helpers.js";
@@ -28,51 +28,7 @@ export const updateNavBar = function () {
 
 /* Adds functionality to the buttons in the navbar */
 export const navBarButtons = function () {
-  // const modal = document.getElementById("formModal");
-
-  // // Login Button
-  // document.getElementById('loginBtn').addEventListener('click', async(event) => {
-  //   modal.style.display = "block";
-    
-  //   const loginForm = new LoginForm();
-  //   document.querySelector('#formModal').innerHTML = await loginForm.getHtml();
-  //   loginForm.afterRender();
-    
-  //   // Define the behavior of the close button after it is in the DOM.
-  //   const span = document.getElementsByClassName("close")[0];
-  //   span.addEventListener('click', () => {
-  //     modal.style.display = "none";
-  //   });
-
-  //   window.onclick = event => {
-  //     if (event.target === modal) {
-  //       modal.style.display = "none";
-  //     }
-  //   };
-  // });
-
-  // // Register Button
-  // document.getElementById('registerBtn').addEventListener('click', async(event) => {
-  //   modal.style.display = "block";
-    
-  //   const registerForm = new RegisterForm();
-  //   document.querySelector('#formModal').innerHTML = await registerForm.getHtml();
-  //   registerForm.afterRender();
-    
-  //   // Define the behavior of the close button after it is in the DOM.
-  //   const span = document.getElementsByClassName("close")[0];
-  //   span.addEventListener('click', () => {
-  //     modal.style.display = "none";
-  //   });
-
-  //   window.onclick = event => {
-  //     if (event.target === modal) {
-  //       modal.style.display = "none";
-  //     }
-  //   };
-  // });
-
-  // Buttons
+  // Buttons with Modal Forms
   new Modal('registerBtn', new RegisterForm());
   new Modal('loginBtn', new LoginForm());
   new Modal('userBtn', new UserForm());
@@ -85,16 +41,4 @@ export const navBarButtons = function () {
     localStorage.removeItem('profile_username');
     navigateTo('/');
   });
-
-  // User Button
-
-  // document.getElementById('userBtn').addEventListener('click', async(event) => {
-  //   const userBtn = document.getElementById('userName')
-
-  //   console.log('user');
-  //   userBtn.textContent = 'Clicked!';
-  //   setTimeout(() => {
-  //     userBtn.textContent = localStorage.getItem('profile_username');
-  //   }, 2000);
-  // });
 };
